+++ pkg/services/object/server.go
@@ -231,15 +231,17 @@ func (s *Server) makeResponseMetaHeader(st *protostatus.Status) *protosession.Re
 	}
 }
 
-func (s *Server) sendPutResponse(stream protoobject.ObjectService_PutServer, resp *protoobject.PutResponse) error {
+func (s *Server) sendPutResponse(stream protoobject.ObjectService_PutServer, resp *protoobject.PutResponse, err error) error {
+	if err != nil {
+		resp.MetaHeader = s.makeResponseMetaHeader(util.ToStatus(err))
+	}
+
 	resp.VerifyHeader = util.SignResponse(&s.signer, resp)
 	return stream.SendAndClose(resp)
 }
 
 func (s *Server) sendStatusPutResponse(stream protoobject.ObjectService_PutServer, err error) error {
-	return s.sendPutResponse(stream, &protoobject.PutResponse{
-		MetaHeader: s.makeResponseMetaHeader(util.ToStatus(err)),
-	})
+	return s.sendPutResponse(stream, new(protoobject.PutResponse), err)
 }
 
 type putStream struct {
@@ -424,11 +426,7 @@ func (s *Server) Put(gStream protoobject.ObjectService_PutServer) error {
 		if req, err = gStream.Recv(); err != nil {
 			if errors.Is(err, io.EOF) {
 				resp, err = ps.close()
-				if err != nil {
-					return s.sendStatusPutResponse(gStream, err)
-				}
-
-				err = s.sendPutResponse(gStream, resp)
+				err = s.sendPutResponse(gStream, resp, err)
 				return err
 			}
 			return err
@@ -482,15 +480,16 @@ func (s *Server) Put(gStream protoobject.ObjectService_PutServer) error {
 	}
 }
 
-func (s *Server) signDeleteResponse(resp *protoobject.DeleteResponse) *protoobject.DeleteResponse {
+func (s *Server) signDeleteResponse(resp *protoobject.DeleteResponse, err error) *protoobject.DeleteResponse {
+	if err != nil {
+		resp.MetaHeader = s.makeResponseMetaHeader(util.ToStatus(err))
+	}
 	resp.VerifyHeader = util.SignResponse(&s.signer, resp)
 	return resp
 }
 
 func (s *Server) makeStatusDeleteResponse(err error) *protoobject.DeleteResponse {
-	return s.signDeleteResponse(&protoobject.DeleteResponse{
-		MetaHeader: s.makeResponseMetaHeader(util.ToStatus(err)),
-	})
+	return s.signDeleteResponse(new(protoobject.DeleteResponse), err)
 }
 
 type deleteResponseBody protoobject.DeleteResponse_Body
@@ -564,11 +563,11 @@ func (s *Server) Delete(ctx context.Context, req *protoobject.DeleteRequest) (*p
 	p.WithAddress(addr)
 	p.WithTombstoneAddressTarget((*deleteResponseBody)(&rb))
 	err = s.handlers.Delete(ctx, p)
-	if err != nil {
+	if err != nil && !errors.Is(err, apistatus.ErrIncomplete) {
 		return s.makeStatusDeleteResponse(err), nil
 	}
 
-	return s.signDeleteResponse(&protoobject.DeleteResponse{Body: &rb}), nil
+	return s.signDeleteResponse(&protoobject.DeleteResponse{Body: &rb}, err), nil
 }
 
 func (s *Server) signHeadResponse(resp *protoobject.HeadResponse, sign bool) *protoobject.HeadResponse {
@@ -1873,15 +1872,17 @@ func (s *Server) Replicate(_ context.Context, req *protoobject.ReplicateRequest)
 	return resp, nil
 }
 
-func (s *Server) signSearchResponse(resp *protoobject.SearchV2Response) *protoobject.SearchV2Response {
+func (s *Server) signSearchResponse(body *protoobject.SearchV2Response_Body, err error) *protoobject.SearchV2Response {
+	var resp = new(protoobject.SearchV2Response)
+
+	if err != nil {
+		resp.MetaHeader = s.makeResponseMetaHeader(apistatus.FromError(err))
+	}
+	if err == nil || errors.Is(err, apistatus.ErrIncomplete) {
+		resp.Body = body
+	}
 	resp.VerifyHeader = util.SignResponse(&s.signer, resp)
 	return resp
-}
-
-func (s *Server) makeStatusSearchResponse(err error) *protoobject.SearchV2Response {
-	return s.signSearchResponse(&protoobject.SearchV2Response{
-		MetaHeader: s.makeResponseMetaHeader(apistatus.FromError(err)),
-	})
 }
 
 func (s *Server) SearchV2(ctx context.Context, req *protoobject.SearchV2Request) (*protoobject.SearchV2Response, error) {
@@ -1891,11 +1892,11 @@ func (s *Server) SearchV2(ctx context.Context, req *protoobject.SearchV2Request)
 	)
 	defer s.pushOpExecResult(stat.MethodObjectSearchV2, err, t)
 	if err = icrypto.VerifyRequestSignaturesN3(req, s.fsChain); err != nil {
-		return s.makeStatusSearchResponse(err), nil
+		return s.signSearchResponse(nil, err), nil
 	}
 
 	if s.fsChain.LocalNodeUnderMaintenance() {
-		return s.makeStatusSearchResponse(apistatus.ErrNodeUnderMaintenance), nil
+		return s.signSearchResponse(nil, apistatus.ErrNodeUnderMaintenance), nil
 	}
 
 	reqInfo, err := s.reqInfoProc.SearchV2RequestToInfo(req)
@@ -1905,23 +1906,19 @@ func (s *Server) SearchV2(ctx context.Context, req *protoobject.SearchV2Request)
 			bad.SetMessage(err.Error())
 			err = bad // defer
 		}
-		return s.makeStatusSearchResponse(err), nil
+		return s.signSearchResponse(nil, err), nil
 	}
 	if !s.aclChecker.CheckBasicACL(reqInfo) {
 		err = basicACLErr(reqInfo) // needed for defer
-		return s.makeStatusSearchResponse(err), nil
+		return s.signSearchResponse(nil, err), nil
 	}
 	err = s.aclChecker.CheckEACL(req, reqInfo)
 	if err != nil && !errors.Is(err, aclsvc.ErrNotMatched) { // Not matched -> follow basic ACL.
 		err = eACLErr(reqInfo, err)
-		return s.makeStatusSearchResponse(err), nil
+		return s.signSearchResponse(nil, err), nil
 	}
 
-	body, err := s.processSearchRequest(ctx, req)
-	if err != nil {
-		return s.makeStatusSearchResponse(err), nil
-	}
-	return s.signSearchResponse(&protoobject.SearchV2Response{Body: body}), nil
+	return s.signSearchResponse(s.processSearchRequest(ctx, req)), nil
 }
 
 func verifySearchFilter(f *protoobject.SearchFilter) error {
@@ -1981,7 +1978,7 @@ func (s *Server) processSearchRequest(ctx context.Context, req *protoobject.Sear
 	}
 
 	res, newCursor, err := s.ProcessSearch(ctx, req)
-	if err != nil {
+	if err != nil && !errors.Is(err, apistatus.ErrIncomplete) {
 		return nil, err
 	}
 
@@ -1997,7 +1994,7 @@ func (s *Server) processSearchRequest(ctx context.Context, req *protoobject.Sear
 	if newCursor != nil {
 		resBody.Cursor = base64.StdEncoding.EncodeToString(newCursor)
 	}
-	return resBody, nil
+	return resBody, err
 }
 
 func (s *Server) ProcessSearch(ctx context.Context, req *protoobject.SearchV2Request) ([]sdkclient.SearchResultItem, []byte, error) {